requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.0
Flask-Mail==0.9.1
bcrypt==4.1.3
Pillow==11.0.0

